refactor(frontend): improved api services file structure
