+++ agenta-web/src/components/EvaluationTable/SingleModelEvaluationTable.tsx
@@ -16,7 +16,7 @@ import {
     message,
 } from "antd"
 import {callVariant} from "@/services/api"
-import {updateEvaluationScenario, updateEvaluation} from "@/services/human-evaluations"
+import {updateEvaluationScenario, updateEvaluation} from "@/services/human-evaluations/api"
 import {useVariants} from "@/lib/hooks/useVariant"
 import {useRouter} from "next/router"
 import {EvaluationFlow} from "@/lib/enums"
