+++ agenta-web/src/components/HumanEvaluationModal/HumanEvaluationModal.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from "react"
 import {GenericObject, JSSTheme, Parameter, Variant} from "@/lib/Types"
 import {fetchVariants} from "@/services/api"
-import {createNewEvaluation} from "@/services/human-evaluations"
+import {createNewEvaluation} from "@/services/human-evaluations/api"
 import {isDemo} from "@/lib/helpers/utils"
 import {Button, Col, Dropdown, MenuProps, Modal, ModalProps, Row, Spin, message} from "antd"
 import {getErrorMessage} from "@/lib/helpers/errorHandler"
@@ -14,7 +14,7 @@ import {useAppTheme} from "../Layout/ThemeContextProvider"
 import {createUseStyles} from "react-jss"
 import EvaluationErrorModal from "../Evaluations/EvaluationErrorModal"
 import {dynamicComponent} from "@/lib/helpers/dynamic"
-import {useLoadTestsetsList} from "@/services/testsets"
+import {useLoadTestsetsList} from "@/services/testsets/api"
 
 type StyleProps = {
     themeMode: "dark" | "light"
