+++ agenta-web/src/components/SaveTestsetModal/SaveTestsetModal.tsx
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from "react"
 import {useAppId} from "@/hooks/useAppId"
 import {Evaluation, EvaluationScenario} from "@/lib/Types"
 import {EvaluationFlow} from "@/lib/enums"
-import {createNewTestset} from "@/services/testsets"
+import {createNewTestset} from "@/services/testsets/api"
 import {Form, Input, Modal} from "antd"
 
 type EvaluationRow = EvaluationScenario & {
