+++ agenta-web/src/components/TestSetTable/TestsetTable.tsx
@@ -5,7 +5,7 @@ import {Button, Input, Tooltip, Typography, message} from "antd"
 import TestsetMusHaveNameModal from "./InsertTestsetNameModal"
 import {DeleteOutlined, EditOutlined, PlusOutlined} from "@ant-design/icons"
 import {fetchVariants} from "@/services/api"
-import {createNewTestset, fetchTestset, updateTestset} from "@/services/testsets"
+import {createNewTestset, fetchTestset, updateTestset} from "@/services/testsets/api"
 import {useRouter} from "next/router"
 import {useAppTheme} from "../Layout/ThemeContextProvider"
 import useBlockNavigation from "@/hooks/useBlockNavigation"
