+++ agenta-web/src/components/pages/evaluations/evaluators/EvaluatorCard.tsx
@@ -6,7 +6,7 @@ import {createUseStyles} from "react-jss"
 import dayjs from "dayjs"
 import Image from "next/image"
 import AlertPopup from "@/components/AlertPopup/AlertPopup"
-import {deleteEvaluatorConfig} from "@/services/evaluations"
+import {deleteEvaluatorConfig} from "@/services/evaluations/api"
 import {useAtom} from "jotai"
 import {evaluatorsAtom} from "@/lib/atoms/evaluation"
 import {checkIfResourceValidForDeletion} from "@/lib/helpers/evaluate"
