+++ agenta-web/src/components/pages/evaluations/evaluators/Evaluators.tsx
@@ -5,7 +5,7 @@ import {Button, Empty, Input, Space, Spin} from "antd"
 import {PlusCircleOutlined} from "@ant-design/icons"
 import NewEvaluatorModal from "./NewEvaluatorModal"
 import {useAppId} from "@/hooks/useAppId"
-import {fetchAllEvaluatorConfigs} from "@/services/evaluations"
+import {fetchAllEvaluatorConfigs} from "@/services/evaluations/api"
 import {useAtom} from "jotai"
 import {evaluatorConfigsAtom} from "@/lib/atoms/evaluation"
 import {JSSTheme} from "@/lib/Types"
