+++ agenta-web/src/pages/apps/[app_id]/annotations/human_a_b_testing/[evaluation_id]/index.tsx
@@ -1,7 +1,10 @@
 import ABTestingEvaluationTable from "@/components/EvaluationTable/ABTestingEvaluationTable"
 import type {Evaluation} from "@/lib/Types"
-import {fetchLoadEvaluation, fetchAllLoadEvaluationsScenarios} from "@/services/human-evaluations"
-import {fetchTestset} from "@/services/testsets"
+import {
+    fetchLoadEvaluation,
+    fetchAllLoadEvaluationsScenarios,
+} from "@/services/human-evaluations/api"
+import {fetchTestset} from "@/services/testsets/api"
 import {useRouter} from "next/router"
 import {useEffect, useState} from "react"
 import {fetchVariants} from "@/services/api"
