+++ agenta-web/src/pages/apps/[app_id]/endpoints/index.tsx
@@ -11,7 +11,7 @@ import {isDemo} from "@/lib/helpers/utils"
 import {dynamicComponent} from "@/lib/helpers/dynamic"
 import {useVariant} from "@/lib/hooks/useVariant"
 import {fetchVariants, fetchAppContainerURL} from "@/services/api"
-import {fetchEnvironments} from "@/services/deployment"
+import {fetchEnvironments} from "@/services/deployment/api"
 import {ApiOutlined, AppstoreOutlined, HistoryOutlined} from "@ant-design/icons"
 import {Alert, Collapse, CollapseProps, Empty, Radio, Tabs, Tooltip, Typography} from "antd"
 import {useRouter} from "next/router"
