+++ agenta-web/src/pages/apps/[app_id]/evaluations/new-evaluator.tsx
@@ -1,7 +1,7 @@
 import Evaluators from "@/components/pages/evaluations/evaluators/Evaluators"
 import {useAppId} from "@/hooks/useAppId"
 import {evaluatorConfigsAtom, evaluatorsAtom} from "@/lib/atoms/evaluation"
-import {fetchAllEvaluatorConfigs, fetchAllEvaluators} from "@/services/evaluations"
+import {fetchAllEvaluatorConfigs, fetchAllEvaluators} from "@/services/evaluations/api"
 import {useAtom} from "jotai"
 import React, {useEffect} from "react"
